[#139340] Add ability to export account search results to CSV (#2155)
# Release Notes

Add ability to export account search results to CSV.

# Additional Context

I chose not to send this asynchronously via email as I suspect this won't be slow enough to have a problem. I've asked the client. We can always switch it over later.

I extracted the `AccountSearcher` to help clean up the controller, but with an eye towards Dartmouth's upcoming feature of being able to search by more fields for accounts. I chose not to do any further refactoring there in favor of doing the refactoring when we get to that story.