[#148320] Improve performance of relay statuses (#2198)
# Release Notes

Improves the performance of relay status fetching from the reservation timeline view.

* Only query a relay once based on IP/port/outlet
* Avoid unnecessary database lookups and inserts
* Do not return non-physical relays as they do not display on the timeline view.

It also adds some logging to the network queries.

# Additional Context

I'm not 100% we need the `InstrumentStatus`es stored in the database. I suspect there was an intention of using it as a kind of cache, but it's not really used anywhere. I feel like the entire model/database table could be removed, but I think that's beyond the scope of this PR. Values are still inserted when switching an instrument on and off.

Another improvement that would be possible is to only query by IP/port so if a relay has two outlets we only query it once.