fix bug in chart_of_accounts caused by validation triggering on .save!
The problem was this:
- We have an existing account owner in DB, but out Chart of Accounts
feed tells us we have a new account owner
- So we set :deleted_at on the existing owner to remove it
- Then we add the new owner

However, the problem was that calling .save! on the existing owner
record would trigger validations, which would throw an exception because
there must be an account owner.

By using only .touch we avoid the model validation being run before we
add the new account owner. When we do add the new owner, we call .save!
and thus the validation is run at this point. Since the validation is
run before this function returns, I believe skipping the validation
in the first step is okay.