Tech task: Fix spurious spec failure (#2241)
# Release Notes

Tech task: Fix spurious test failure.

# Additional Context

```
Failure/Error: click_link "Clone Payment Source Memberships"

Capybara::Ambiguous:
  Ambiguous match, found 2 elements matching visible link "Clone Payment Source Memberships"

./vendor/bundle/ruby/2.5.0/gems/capybara-3.31.0/lib/capybara/node/finders.rb:300:in `block in synced_resolve'
./vendor/bundle/ruby/2.5.0/gems/capybara-3.31.0/lib/capybara/node/base.rb:83:in `synchronize'
./vendor/bundle/ruby/2.5.0/gems/capybara-3.31.0/lib/capybara/node/finders.rb:291:in `synced_resolve'
./vendor/bundle/ruby/2.5.0/gems/capybara-3.31.0/lib/capybara/node/finders.rb:52:in `find'
./vendor/bundle/ruby/2.5.0/gems/capybara-3.31.0/lib/capybara/node/actions.rb:42:in `click_link'
./vendor/bundle/ruby/2.5.0/gems/capybara-3.31.0/lib/capybara/session.rb:759:in `block (2 levels) in <class:Session>'
./vendor/bundle/ruby/2.5.0/gems/capybara-3.31.0/lib/capybara/dsl.rb:58:in `block (2 levels) in <module:DSL>'
./spec/system/admin/clones_account_memberships_spec.rb:37:in `block (3 levels) in <top (required)>'
```

This was super tricky to figure out. What it came down to is we have a couple users whose sequence of name coming out of the factory is out of sync with the sequence of email addresses. I think this happens where if you explicitly set the value in a `FactoryBot.create` it doesn't increment the sequence, and I know we do that in a few places.

So we had one user with name "User 724" and email "user722@example.com" and then another user with "User 726" and "user724@example.com". Because our user search is pretty generous (in order to support first name, last name, and full name searching), a search for "User 724" matches the first user's name and the second user's email address. 

That's why we were seeing two results come back which was then confusing capybara. I figured this out by throwing this code in there so I could actually see what was happening on the failures on CI. There's probably a way to do `save_and_open_page` on CI, but I don't know how.

```ruby
unless page.has_link?("Clone Payment Source Memberships", count: 1)
  raise "Too many links for user: #{original_user.name} -- All: #{User.all.map(&:name) }\n#{page.body}"
end
```

The detail is the query gets converted to `%user%724%` and then matched against the various columns. In general I'm ok with it being as lenient as it is and no one's complained about it. Better for the search to return more results than to not return a result.