[#153460] Fix error on journal creation with split account (#2285)
# Release Notes

Fixes an error when creating a journal containing a split account.

# Additional Context

The error is: `Mysql2::Error: Duplicate entry '148655' for key 'PRIMARY'`

This came up at Dartmouth. I was able to reproduce it in open, but struggled to generate a test case that failed. It seemed to be happening when the purchaser of the order detail was a member of all of the subaccounts (this may be coincidental).

What is happening is for split accounts, when you create a journal we generate virtual order details containing the split data (which account, cost, etc). We then generate a journal row for each of these virtual order details. In some cases, when we are saving the journal row to the database, Rails is also trying to auto-save the virtual order detail. Since the virtual order detail has the same ID as the actual order detail, we are unable to save it because of the duplicate ID.

This autosaving does not happen all the time. And we've never seen this error at NU. So it's not clear what triggers the autosave.

As I said, I was unable to write a test case that triggers the error, but the fix worked in my testing in the dev environment.