[#153552] Fix issue where relays were getting deleted (#2301)
# Release Notes

Fix situation where when creating a new product group containing an instrument with a relay the instrument was getting set to "Reservation Only"

# Additional Context

This seems to be happening only when creating a new product group. Adding to an existing group does not trigger the bug.

I'm actually a little surprised this hasn't bitten us before in some other case like bundles.

What's happening is that the validations on `Product`/`Instrument` are being triggered as part of the `ProductDisplayGroup`'s validation. I can't explain why it's only being triggered in this single case. But because we're triggering the validations, there's some code in there that should only be triggered during product edit that deletes and rebuilds the relay. However, since we're only running the validations and not the save, the new relay is not persisted.

I ended up doing two things to fix it:

* At the base level, only run the destroy/create relay callback during the product edit/update process. This should avoid us getting bitten by this in other associations we may create at some point.

* Don't validate the products from the display groups. In this case, we should be safe to assume the products are already valid.