[#153049] Fix export raw link (#2321)
# Release Notes

Fixes the export raw link that was somewhat broken in #2317. 

# Additional Context

Because the controller changed to use a different parameter, whatever you put in the dialog prompt was getting ignored and it was always falling back to the current user's email address.

I considered trying to refactor it use all the newer shared code in `export_csv_report.coffee`, but this code is doing a lot of extra stuff like munging query parameters and it seemed like more effort than it was worth. I did do some renaming so it uses our more modern `js--` class prefix rather than IDs.