[#153781] Fixes for oracle full text search (#2332)
# Release Notes

Fixes a hard error in product searching in oracle when the query contains only curly braces. 

# Additional Context

We were handling the empty string case before, but once we started removing the braces, it would pass the empty string check, but then become empty. Oracle doesn't like those kinds of queries.

I'll have a PR in NU with this same code in NU to validate the specs against oracle.