[#154126] Refactor welcome email (#2362)
# Release Notes

Tech task: Refactor new user welcome email

# Additional Context

I wanted to do this as part of the updating of the welcome email for UMass. There was a lot of duplication that we don't need between the text version and the html version. This will require a little bit of cleanup in the other schools but it should be straightforward, plus I improved the mailer preview so there's separate previews for internal vs external users.

One last thing I did was renamed `BaseMailer` to `ApplicationMailer`, which is the standard name for the mailer superclass for Rails (we originally created these classes before Rails had a standard superclass).