Tech task: Move google analytics key out of constant into settings (#2371)
# Release Notes

Tech task: Move google analytics key out of per-environment constant into Settings.

# Additional Context

NU is the only one that uses this as far as I know. My plan is to open a new PR in nucore-nu that adds the key to `config/settings/production.yml` so that it's prepped for when this comes in.