Tech task: Refactor 404/403 handling for non-HTML pages (#2380)
# Release Notes

Tech task: Refactor 404/403 handling for non-HTML pages. 

# Additional Context

The changed code deals with 404 not founds and 403 unauthorized for things like PDFs (statements) and ICS calendar invites to have them display as an HTML error page. For example, if you mess with the ID of a statement, then you should see a 404 page, even if the extension is `.pdf`. Without this handling, we would get template not found errors.

The code as it was broke in Rails 6. This change will work in both Rails 5.2 and 6.0.