Properly handle commas in facility names when setting the "from" name in bulk emails (#2384)
# Release Notes

Fix issue in bulk email when the facility has a comma in the name.

# Additional Context

Dartmouth was having deliveries get blocked because they would see multiple "From" addresses in the email header. This was being caused because of how the comma was being interpreted. With this change, the from header will have quotes where necessary to probably escape everything.

This was from the facility "Electron Microscopy - Irradiation, Pre-clinical Imaging and Microscopy Resource".

```
From: <Electron.Microscopy.-.Irradiation@radar-app01.dartmouth.edu>,
Pre-clinical Imaging and Microscopy Resource
<Cancer.Center.Core.Operations@Dartmouth.edu>
```