Update validation so it works properly with expense_accounts feature disabled (#2385)
# Release Notes

Tech task: Update `Product#account` validation so it works properly in tests when the feature is disabled.

# Additional Context

Originally, this validation would never be added at all if the `expense_accounts` feature were disabled (like it is a UConn, and I'm guessing _was_ at UIC). Then, when we turned on the feature in the specs (see `spec/models/product_spec.rb:172`), the validation would not be run because it's not there. Now it will conditionally run based on the setting at runtime rather than compile/load time.