TECH TASK: Use exact_true option instead of exact (#2389)
# Release Notes

There were warnings in the test logs:
```
The :exact option only has an effect on queries using the XPath#is method. Using it with the query "th" has no effect.
The :exact option only has an effect on queries using the XPath#is method. Using it with the query "td" has no effect.
```
... and in fact the `:exact` option had no effect.  For example, `expect(page).to have_selector("td", text: "$1.0", exact: true)` passes unexpectedly.  

This change restores expected behavior and resolves the log noise.