Tech task: Update default PDF paper size to letter (#2395)
# Release Notes

Tech task: update default paper size in PDFs to US Letter.

# Additional Context

When we upgraded the Prawn gems in #2377, that changed the default paper size to A4. Fortunately, CI caught some errors in NU because one of the tables wasn't wide enough in the new paper size.

The documentation [suggests putting this configuration into an initializer](https://github.com/cortiz/prawn-rails#default-configuration), but this is the only place we make PDFs, this seems to be a more obvious spot a developer might see to change it if they're working on making a custom statement. (A4 is a very standard paper size outside the US).