Restructure uploaded file paths and improve instructions for migrating to S3 (#2365)
# Release Notes

Tech task: Update the file paths for uploaded files and improve instructions and helper rake takes to aid in migrating to S3.

# Breaking changes

The setting for where uploaded files are stored locally has changed from `public/files` to `public/system/:class`. You either need to change it back to its previous value or run the rake task as part of your deployment (CHANGELOG.md).

In addition, if you are using S3 already, the path for the keys should now just be `paperclip.aws_access_key` and `paperclip.aws_secret_access_key`.

# Notes

We'll need to update the access keys for NU, but their file paths in S3 already have the `:class` interpolation variable, so nothing should need to move for them.