[#154764] Fix issue where quantity was not being updated (#2426)
# Release Notes

Fix an issue when ordering on behalf if you update the quantity field, the value of quantity nor note and reference id are not persisted.

# Additional Context

To replicate:
* Add an item with a note to the cart
* Fill in the note and reference id fields
* Change the quantity field
* Click "Purchase"
* You will see the "Quantities have change. Please review" message, but the note, quantity, and reference fields all revert to their previous values.
* Clicking Update does the same thing

It always reverted to it's previous vlue:
Put in 3, click update -> reverts to 1
Put in 4, click update -> field shows 3

## Tech notes

While I had no problem replicating this in the browser, I struggled to get a failing spec. I'm not sure what the difference in behavior is. I left in the specs, but they were still passing before the other changes. I eventually gave up.

Ultimately, the issue is the order details getting updated are not the same objects as those that were getting displayed. So when you submitted the form, the it would read the database values into one set of objects, update another, and then update the first set of objected.

Both of the changes: the `inverse_of` and the `select` fixed the issue on its own, but it felt like both were reasonable things to put in.