Fix time range display issue where times are the same day of the month (#2427)
# Release Notes

Fix time range display when times are one month apart on the same day of month (e.g. 7/1 and 8/1).

# Additional Context

They were being treated as if they were on the same day, so the range was appearing as `Fri, 07/01/2016 12:00 PM - 12:00 PM`.

I happened to notice the pending test, so I figured I'd take a quick look into this.