Tech Task: Remove temporary logging (#2554)
# Release Notes

This logging was added as part of #2497.  The problem there was that if a user enters "0" for the actual duration, the reservation would silently fail validations and the order detail would move out of the problem queue (which it shouldn't).  I added this logging to try to understand why reservations are failing, but I don't think this provides much value since #2497 prevents silent failures.