Clean up spec
This spec file was using travel_back to return to system time (not the global time lock).
Then it was also using the :time_travel meta flag to set the time to 12:30 on the current day.
The specs work fine on system time, so just calling travel_back is sufficient, and saves the
complexity of understanding :time_travel.