[#159023] Only apply bulk reconcile note when the bulk note checkbox has been checked (#2812)
# Release Notes
This PR adds logic to account for the bulk note checkbox value when determining how to update order detail records.  
Also changes the bulk note checkbox checked value from `"0"` to `"1"`.  The value is not sent in params when unchecked.

Handles the following edge case:
- User checks the bulk note checkbox
- User enters a bulk note
- User unchecks the bulk note checkbox
- User reconciles transactions

In this case the bulk note will not be applied, which is the expected behavior.