[158844] complete rails upgrade to 6.0 (#2952)
# Release Notes

This removes 5.0 defaults and loads 6.0 defaults.

The vast majority of the work here is setting up classes and modules to be autoloaded from files and directories in the [new zeitwerk's way](https://guides.rubyonrails.org/classic_to_zeitwerk_howto.html). This mostly means, let the module structure match the folder structure and let the class (or module) name match the file name.

Additionally most things were moved out of the `lib/` directory because that's not standard to be autoloaded and some of the files in there are more correctly put in a `concerns/` directory.

Zeitwerk takes care of loading things from the `app` directory, so that no longer needs to be configured in `application.rb`, unless you want a non-standard root, as is the case with the `external_services` directory.