Pulls in style changes from OSU (#3005)
Some schools are interested in more flexibility regarding the branding/customization of the header.

This adds some settings.yml and sass variables to make the site header more customizable, particularly regarding the logo. The dimensions of the header can now be changed. It is also possible to elect not to show the app name beside the header.