[#160767] Adds NetID as a term in the translation file (#3145)
# Release Notes

Not every school calls their online SSO account "NetID". This reorganizes the translation files so that "NetID" is easy to change to something else.

The order import template contains "Netid" and is a static file. So this also adds `order_import_template_name` to the `settngs.yml` file, in case a school would like to change to its own template with school specific language.