Disables class caching in the development environment (#3078)
# Release Notes

Cached classes requires a lot of server restarts when developing. This disables class caching.

It also adjusts some loading behavior of the `C2po` engine so the application works correctly with respect to the caching. The way `C2po` was setup to load, `AccountConfig`'s reconcilable accounts would disappear. This make the reconcilable account code a bit more clear.