[#160368] Add image processing to DownloadableFiles::Image (#3221)
Users can attached image files of any size and dimension to a Facility, which can make the `facility_tile_list` grid layout in unexpected ways. This uses `ActiveStroage#variant` to make images the same size and dimension, it adds colored padding if the image is the wrong aspect ratio.

This depends on `libvips` to do the image processing. That or `ImageMagick` must be installed on the server for this to work.