[#16088] SciShield adapter for safety certs (#3225)
This adds an adapter and client class for integrating NUCore with SciShield for the purpose of checking the status of research safety certifications. We now have 2 schools using SciShield (formerly Bioraft) so this makes it easy to share the logic. Credentials and API endpoint are configurable via ENV variables. Also, this makes it possible to specify an adapter class to use for research safety lookups via ENV variable. This way we have more control over which adapter is used in production without needing to do a separate release.