Adds rake task to add template_file to order detail
In the case that a service is modified to require a form after it has been
purchased, the order detail will state that it's missing a form, but there's
no way to add the needed template_result file via the UI after purchase.
This task adds a template_result file to an order detail by copying the template
file from the product that was purchased.