[#161053] Pulls in affiliate fixes from OSU (#3279)
# Release Notes

When the change was pulled into OSU, spec failures occurred. This cherry picks the fixes for those failures.

It adjusts how the `po_require_affiliate_account` feature flag works. It removes the `affiliate_id` validation instead of removing the `AffiliateAccount` include. And it insures the validation test only runs for `PurchaseOrderAccount`