Tech task - Add Dockerfile for Github Actions testing (#3342)
# Release Notes

As we prepare to move away from CircleCI this will be helpful to have here in nucore-open.
I used [prograils' ruby image](https://github.com/prograils/docker-ruby-node-chrome-pack/blob/2.7.5/Dockerfile) as a base and added in ENV variables to control node version and bundler version.  This allows CI testing to run in an environment that matches production more closely.