Removes date conversion on `orderedAtDate` key (#3422)
# Release Notes

This removes the `to_date` conversion on the `OderDetails`'s `created_at`.

Converting this from a DateTime to a Date seems to result in form.io turning it back into a DateTime, probably at 12:00 AM UTC of the given date. Then converting it into to CST, resulting in the date being yesterday rather than today.