Add a type check to make sure we don't get confused in the future
When no transactions were checked, we were passing `moment` an `undefined` value for `earliestFulfilledAtDate`.
It seems to default to the current day in that case.
Although we only expect this odd situation to happen in the context of specs
running in the "global time lock" date, it seems good to be explicit about
the behavior we want when no transactions are checked.