[#161859] Moves nonbillable user creation and update its attributes  (#3645)
# Release Notes

Creating the user who owns the Nonbillable account is more at home in the NonbillableAccount class than it is in the User class.

Some schools have users with the same email address as the support email address, which means the nonbillable user can't be created. This changes the email address the nonbillable user uses.