Remove unnecessary account_price_group_member's and related complexity
We shouldn't be assigning accounts to price groups manually here.
This is left over from a previous iteration of billing_mode functionality,
before we were creating price rules automatically.

It's interesting to know that it is apparently possible
for an account to belong to both the base and external price groups.
We don't expect that to happen in practice though.