Use a single price policy for Nonbillable products
It's simpler to reason about Nonbillable this way.
We set the price group explicitly everywhere that matters,
so there should be no need to have any more than a single price policy.
I'd also prefer to live with the duplication for now,
in case there are further changes in the near future.