[#161966] Don't auto-reconcile problem orders (#3683)
# Release Notes

Nonbillable and Skip Review transactions should go to the problem queue like everyone else.

Before this change, Nonbillable and Skip Review transactions would get marked "Complete" as part of being moved to the problem queue, and then go automatically to "Reconciled". The system does not expect the order detail to be reconciled, so it raises an error and rolls everything back:
`raise "Trying to move Order ##{@order_detail} to problem queue, but it's not a problem" unless @order_detail.problem?`

The code's definition of "problem" is more complex than it needs to be but it requires the the order detail to be in the "complete" state:
```
  before_save :set_problem_order
  def set_problem_order
    self.problem = problem_description_key.present?
    update_fulfilled_at_on_resolve if time_data.present?
    true # problem might be false; we need the callback chain to continue
  end
# ...
  def problem_description_key
    Array(problem_description_keys).first
  end

  def problem_description_keys
    return [] unless complete?

    time_data_problem_key = time_data.problem_description_key
    price_policy_problem_key = :missing_price_policy if price_policy.blank?
    missing_form_problem_key = :missing_form if missing_form?

    [time_data_problem_key, price_policy_problem_key, missing_form_problem_key].compact
  end
```

To fix this, we just need to prevent auto-reconciling.  This allows `problem_description_keys` to return the expected values, `problem_description_key.present?` will return truthy, and `problem` will be set to `true`.